feat: refact update

+++ src/app/home/pages/user-page/user-page.component.ts
@@ -116,16 +116,22 @@ export class UserPageComponent implements OnInit {
 
 
 upImage(){
+
+}
+
+UploadImageFacto(){
   if(!this.file) return
+  if(!this.currentUser) return
   const id = this.currentUser?._id
+  const nombre = this.currentUser?.name
   console.log(id)
-  this.locationService.uploadImageUser(this.file).subscribe( (response) => {
-    const objectTest = {image: response.secure_url}
+  this.locationService.uploadImageFacto(this.file, nombre).subscribe( (response) => {
+    const objectTest = {image: response.secure_url, name: 'usuario'}
     if(!id) return
     this.deleteImageBeforeUpdate(id);
-  this.locationService.updatedLocationImageUser(id, objectTest).subscribe( ()=> {
+  this.locationService.updatedLocationImageUserRefact(id, objectTest).subscribe( ()=> {
     if(!this.currentUser?.image) return
-    this.currentUser.image = objectTest.image
+      this.currentUser.image = objectTest.image
     })
     
   })
